# taskify-front/assets/openapi.json
{
    "openapi": "3.1.0",
    "info": {
    "title": "Taskify Backend",
    "version": "0.1.0"
    },
    "paths": {
...
    "/api/v1/lists/board/{board_id}": {
    "post": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Create List In Board",
    "operationId": "ListController_create_list_in_board_api_v1_lists_board__board_id__post",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ListCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ListRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Get Board Lists",
    "operationId": "ListController_get_board_lists_api_v1_lists_board__board_id__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_ListRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/lists/{list_id}": {
    "delete": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Delete List",
    "operationId": "ListController_delete_list_api_v1_lists__list_id__delete",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "list_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "List Id"
    }
    }
    ],
    "responses": {
    "204": {
    "description": "Successful Response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Get List",
    "operationId": "ListController_get_list_api_v1_lists__list_id__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "list_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "List Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ListRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Patch List",
    "operationId": "ListController_patch_list_api_v1_lists__list_id__patch",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "list_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "List Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ListUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ListRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/lists/board/public/{board_hash}": {
    "get": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Get Public Lists",
    "operationId": "ListController_get_public_lists_api_v1_lists_board_public__board_hash__get",
    "parameters": [
    {
    "name": "board_hash",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "title": "Board Hash"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_ListRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/lists/board/{board_id}/reorder": {
    "post": {
    "tags": [
    "Lists"
    ],
    "summary": "Listcontroller.Reorder List",
    "operationId": "ListController_reorder_list_api_v1_lists_board__board_id__reorder_post",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "board_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "array",
    "items": {
    "type": "string",
    "format": "uuid"
    },
    "title": "List Order"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_ListRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/cards/list/{list_id}": {
    "post": {
    "tags": [
    "Cards"
    ],
    "summary": "Cardcontroller.Create Card",
    "operationId": "CardController_create_card_api_v1_cards_list__list_id__post",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "list_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "List Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CardCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Cards"
    ],
    "summary": "Cardcontroller.Get List Cards",
    "operationId": "CardController_get_list_cards_api_v1_cards_list__list_id__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "list_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "List Id"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_CardRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/cards/{card_id}": {
    "delete": {
    "tags": [
    "Cards"
    ],
    "summary": "Cardcontroller.Delete Card",
    "operationId": "CardController_delete_card_api_v1_cards__card_id__delete",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "card_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Card Id"
    }
    }
    ],
    "responses": {
    "204": {
    "description": "Successful Response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Cards"
    ],
    "summary": "Cardcontroller.Get Card",
    "operationId": "CardController_get_card_api_v1_cards__card_id__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "card_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Card Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "Cards"
    ],
    "summary": "Cardcontroller.Patch Card",
    "operationId": "CardController_patch_card_api_v1_cards__card_id__patch",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "card_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Card Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CardUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/cards/list/{list_id}/public/{board_hash}": {
    "get": {
    "tags": [
    "Cards"
    ],
    "summary": "Cardcontroller.Get Public List Cards",
    "operationId": "CardController_get_public_list_cards_api_v1_cards_list__list_id__public__board_hash__get",
    "parameters": [
    {
    "name": "list_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "List Id"
    }
    },
    {
    "name": "board_hash",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "title": "Board Hash"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_CardRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/cards/{card_id}/move": {
    "post": {
    "tags": [
    "Cards"
    ],
    "summary": "Cardcontroller.Move Card",
    "operationId": "CardController_move_card_api_v1_cards__card_id__move_post",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "card_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Card Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Body_CardController_move_card_api_v1_cards__card_id__move_post"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CardRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/cards/list/{list_id}/reorder": {
    "post": {
    "tags": [
    "Cards"
    ],
    "summary": "Cardcontroller.Reorder Cards",
    "operationId": "CardController_reorder_cards_api_v1_cards_list__list_id__reorder_post",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "list_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "List Id"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "type": "array",
    "items": {
    "type": "string",
    "format": "uuid"
    },
    "title": "Card Order"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_CardRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
...
    "Body_CardController_move_card_api_v1_cards__card_id__move_post": {
    "properties": {
    "target_list_id": {
    "type": "string",
    "format": "uuid",
    "title": "Target List Id"
    },
    "position": {
    "type": "integer",
    "title": "Position"
    }
    },
    "type": "object",
    "required": [
    "target_list_id",
    "position"
    ],
    "title": "Body_CardController_move_card_api_v1_cards__card_id__move_post"
    },
...
    "CardCreate": {
    "properties": {
    "title": {
    "type": "string",
    "title": "Title"
    },
    "description": {
    "type": "string",
    "title": "Description"
    },
    "position": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Position"
    },
    "assigned_to_id": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "Assigned To Id"
    },
    "due_date": {
    "anyOf": [
    {
    "type": "string",
    "format": "date-time"
    },
    {
    "type": "null"
    }
    ],
    "title": "Due Date"
    },
    "is_completed": {
    "type": "boolean",
    "title": "Is Completed",
    "default": false
    }
    },
    "type": "object",
    "required": [
    "title",
    "description"
    ],
    "title": "CardCreate"
    },
    "CardRead": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "created_at": {
    "type": "string",
    "format": "date-time",
    "title": "Created At"
    },
    "updated_at": {
    "type": "string",
    "format": "date-time",
    "title": "Updated At"
    },
    "title": {
    "type": "string",
    "title": "Title"
    },
    "description": {
    "type": "string",
    "title": "Description"
    },
    "position": {
    "type": "integer",
    "title": "Position"
    },
    "list_id": {
    "type": "string",
    "format": "uuid",
    "title": "List Id"
    },
    "assigned_to_id": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "Assigned To Id"
    },
    "due_date": {
    "anyOf": [
    {
    "type": "string",
    "format": "date-time"
    },
    {
    "type": "null"
    }
    ],
    "title": "Due Date"
    },
    "is_completed": {
    "type": "boolean",
    "title": "Is Completed"
    }
    },
    "type": "object",
    "required": [
    "id",
    "created_at",
    "updated_at",
    "title",
    "description",
    "position",
    "list_id",
    "assigned_to_id",
    "due_date",
    "is_completed"
    ],
    "title": "CardRead"
    },
    "CardUpdate": {
    "properties": {
    "title": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Title"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "assigned_to_id": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "Assigned To Id"
    },
    "list_id": {
    "anyOf": [
    {
    "type": "string",
    "format": "uuid"
    },
    {
    "type": "null"
    }
    ],
    "title": "List Id"
    },
    "due_date": {
    "anyOf": [
    {
    "type": "string",
    "format": "date-time"
    },
    {
    "type": "null"
    }
    ],
    "title": "Due Date"
    },
    "is_completed": {
    "anyOf": [
    {
    "type": "boolean"
    },
    {
    "type": "null"
    }
    ],
    "title": "Is Completed"
    }
    },
    "type": "object",
    "title": "CardUpdate"
    },
...
    "HTTPValidationError": {
    "properties": {
    "detail": {
    "items": {
    "$ref": "#/components/schemas/ValidationError"
    },
    "type": "array",
    "title": "Detail"
    }
    },
    "type": "object",
    "title": "HTTPValidationError"
    },
    "ListCreate": {
    "properties": {
    "title": {
    "type": "string",
    "title": "Title"
    },
    "position": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Position"
    }
    },
    "type": "object",
    "required": [
    "title"
    ],
    "title": "ListCreate"
    },
    "ListRead": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "created_at": {
    "type": "string",
    "format": "date-time",
    "title": "Created At"
    },
    "updated_at": {
    "type": "string",
    "format": "date-time",
    "title": "Updated At"
    },
    "title": {
    "type": "string",
    "title": "Title"
    },
    "position": {
    "type": "integer",
    "title": "Position"
    },
    "board_id": {
    "type": "string",
    "format": "uuid",
    "title": "Board Id"
    }
    },
    "type": "object",
    "required": [
    "id",
    "created_at",
    "updated_at",
    "title",
    "position",
    "board_id"
    ],
    "title": "ListRead"
    },
    "ListUpdate": {
    "properties": {
    "title": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Title"
    }
    },
    "type": "object",
    "title": "ListUpdate"
    },
...
    "Page_CardRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/CardRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "page",
    "size"
    ],
    "title": "Page[CardRead]"
    },
    "Page_ListRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/ListRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "page",
    "size"
    ],
    "title": "Page[ListRead]"
    },
...
    "ValidationError": {
    "properties": {
    "loc": {
    "items": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "integer"
    }
    ]
    },
    "type": "array",
    "title": "Location"
    },
    "msg": {
    "type": "string",
    "title": "Message"
    },
    "type": {
    "type": "string",
    "title": "Error Type"
    }
    },
    "type": "object",
    "required": [
    "loc",
    "msg",
    "type"
    ],
    "title": "ValidationError"
    }
    },
    "securitySchemes": {
    "APIKeyCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "taskify_auth"
    }
    }
    }
    }